requirements.txt
# Core & Built-in (no install needed)
# asyncio is part of Python 3.12+ standard library

# Configuration management
PyYAML==6.0.1

# Data processing (fully supports Python 3.12)
pandas==2.2.2
numpy==1.26.4

# Machine learning
scikit-learn==1.5.0

# API requests
requests==2.32.3

# Database ORM
sqlalchemy==2.0.30

# Natural Language Processing (optional)
spacy==3.7.5
nltk==3.8.1
textblob==0.17.1

# Web framework for dashboard (Dash + Plotly)
dash==2.17.1
plotly==5.20.0

# Utilities
python-dateutil==2.9.0.post0  # Updated for Python 3.12 compatibility

# Type hints & validation
typing-extensions==4.12.2
pydantic==2.5.3

# Testing & code quality
pytest==8.2.2
pytest-cov==4.1.0
pytest-asyncio==0.23.2
unittest-xml-reporting==3.2.0
black==24.4.2                 
pylint==3.2.5                 
mypy==1.10.0                  